+++ CONTRIBUTING.md
@@ -81,7 +81,7 @@ https://www.conventionalcommits.org/ or check out the
 3. Make and commit your changes following the
    [commit convention](https://github.com/heroui-inc/heroui/blob/main/CONTRIBUTING.md#commit-convention).
    As you canary, you can run `pnpm build --filter=<module>` and
-   `pnpm test packages/<module>/<pkg>` e.g. `pnpm build --filter=avatar & pnpm test packages/components/avatar` to make sure everything works as expected.
+   `pnpm test packages/<module>/<pkg>` e.g. `pnpm build --filter=@heroui/avatar & pnpm test packages/components/avatar` to make sure everything works as expected.
 
    > To know more about the `--filter` option, please check the turborepo [docs](https://turborepo.org/docs/core-concepts/filtering).
