+++ packages/components/toast/stories/toast.stories.tsx
@@ -17,6 +17,11 @@ import {Toast, ToastProps, ToastProvider, addToast, closeAll} from "../src";
 export default {
   title: "Components/Toast",
   component: Toast,
+  args: {
+    motionProps: {
+      exit: {opacity: 0, y: -300, x: -500, transition: {duration: 1, ease: "easeInOut"}},
+    },
+  },
   argTypes: {
     variant: {
       control: {type: "select"},
@@ -65,6 +70,10 @@ export default {
       control: {
         type: "boolean",
       },
+    },
+    motionProps: {
+      control: "object",
+      description: "Custom properties for Framer Motion",
     },
   },
   decorators: [
